refactor: 프로젝트 구조 개편
useTodoContainer를 pinia store로 변경하면서 todo 모듈 관련 프로젝트 구조를 재정의했습니다.
