+++ src/components/todo/container/TodoContainer.vue
@@ -1,13 +1,13 @@
 <script setup>
-import TodoHeader from "@/components/todo/TodoHeader.vue";
+import TodoContainerHeader from "./TodoContainerHeader.vue";
+import TodoContainerListView from "./TodoContainerListView.vue";
 import TodoInput from "@/components/todo/TodoInput.vue";
-import TodoList from "@/components/todo/TodoList.vue";
 </script>
 
 <template>
   <div class="w-full flex flex-col justify-start items-center gap-1 p-4">
-    <TodoHeader />
-    <TodoList />
+    <TodoContainerHeader />
+    <TodoContainerListView />
     <TodoInput />
   </div>
 </template>
